bugs fixed

+++ routes/web.php
@@ -18,12 +18,12 @@ Route::get('/home', [HomeController::class, 'redirect'])->name('home');
 
 //user routes to book services
 Route::get('/bookservice', [BookserviceController::class, 'bookservice'])->name('user.bookservice')->middleware('auth');
-Route::get('/bookservice/{service}', [BookserviceController::class, 'create'])->name('bookservice.book');
-Route::post('/bookservice/{service}', [BookserviceController::class, 'store'])->name('bookservice.store');
+Route::get('/bookservice/{service}', [BookserviceController::class, 'create'])->name('bookservice.book')->middleware('auth');
+Route::post('/bookservice/{service}', [BookserviceController::class, 'store'])->name('bookservice.store')->middleware('auth');
 
 //user route for bookings details
-Route::get('/bookings', [BookingController::class, 'index'])->name('user.booking');
-Route::delete('/bookings/{id}/cancel', [BookingController::class, 'cancel'])->name('booking.cancel');
+Route::get('/bookings', [BookingController::class, 'index'])->name('user.booking')->middleware('auth');
+Route::delete('/bookings/{id}/cancel', [BookingController::class, 'cancel'])->name('booking.cancel')->middleware('auth');
 
 //admin routes for bookings
 Route::get('/admin/bookings', [BookingController::class, 'showAllBookings'])->name('admin.bookings')->middleware(['auth', 'verified', 'admin']);
@@ -36,7 +36,7 @@ Route::post('/booking/{booking}/assign-staff', 'BookingController@assignStaff')-
 
 
 //resource routes
-Route::resource('staff_view', StaffController::class) ->only(['addview', 'store', 'index', 'create', 'destroy', 'edit'])->middleware(['auth', 'verified', 'admin']);
+Route::resource('staff', StaffController::class) ->only(['addview', 'store', 'index', 'create', 'destroy', 'edit', 'update'])->middleware(['auth', 'verified', 'admin']);
 Route::resource('department', DepartmentController::class)->only(['show', 'index', 'store', 'destroy'])->middleware(['auth', 'verified', 'admin']);
 Route::resource('service', ServiceController::class)->only(['index', 'store', 'show', 'destroy'])->middleware(['auth', 'verified', 'admin']);
 Route::resource('booking', BookingController::class)->middleware(['auth', 'verified', 'admin']);
